fast-response-service/pyproject.toml
[project]
name = "mining-microservice"
version = "0.1.0"
description = "LLM-based microservice for document mining"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3",
    "pandas",
    "python-dotenv",
    "pyjwt",
    "aiohttp",
    "requests",
    "fastapi",
    "uvicorn",
    "python-multipart",
    "mangum"
]
